backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
gradio>=4.0.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
moviepy>=1.0.3
openai>=1.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0
scikit-image>=0.19.0
opencv-python>=4.5.0
numpy>=1.21.0
requests>=2.28.0
gunicorn>=20.1.0

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
coverage>=7.0.0
